+++ src/oxl_utils/subps.py
@@ -1,7 +1,9 @@
 import subprocess
+from time import sleep, time
 from pathlib import Path
 from os import environ, getcwd
 from functools import cache
+from threading import Thread
 
 from .log import log
 
@@ -9,7 +11,7 @@ from .log import log
 def process(
         cmd: (str, list), timeout_sec: int = None, shell: bool = False, timeout_shell: bool = True,
         cwd: Path = None, env: dict = None, env_inherit: bool = False, env_remove: list = None,
-        empty_none: bool = True,
+        empty_none: bool = True, stdin: str = None,
 ) -> dict:
     if cwd is None:
         cwd = getcwd()
@@ -47,10 +49,16 @@ def process(
             shell=shell,
             stdout=subprocess.PIPE,
             stderr=subprocess.PIPE,
+            stdin=subprocess.PIPE,
             cwd=cwd,
             env=env,
         ) as p:
-            b_stdout, b_stderr = p.communicate(timeout=timeout_sec)
+            if stdin is not None:
+                b_stdout, b_stderr = p.communicate(timeout=timeout_sec, input=stdin.encode('utf-8'))
+
+            else:
+                b_stdout, b_stderr = p.communicate(timeout=timeout_sec)
+
             stdout, stderr, rc = b_stdout.decode('utf-8').strip(), b_stderr.decode('utf-8').strip(), p.returncode
 
     except (subprocess.TimeoutExpired, subprocess.SubprocessError, subprocess.CalledProcessError,
@@ -76,10 +84,19 @@ def process(
 def process_with_cache(
         cmd: (str, list), timeout_sec: int = None, shell: bool = False, timeout_shell: bool = True,
         cwd: Path = None, env: dict = None, env_inherit: bool = False, env_remove: list = None,
-        empty_none: bool = True,
+        empty_none: bool = True, stdin: str = None,
 ) -> dict:
     # read-only commands which results can be cached
     return process(
         cmd=cmd, timeout_sec=timeout_sec, shell=shell, timeout_shell=timeout_shell, cwd=cwd,
-        env=env, env_inherit=env_inherit, env_remove=env_remove, empty_none=empty_none,
+        env=env, env_inherit=env_inherit, env_remove=env_remove, empty_none=empty_none, stdin=stdin,
     )
+
+
+def wait_for_threads(threads: list[Thread], timeout: int = 600):
+    break_time = time() + timeout
+    threads_done = False
+
+    while not threads_done and time() < break_time:
+        threads_done = all(not t.is_alive() for t in threads)
+        sleep(0.05)